Add some basic support of non-cartesian CSes.
Implement different calculation of drawing AABB, for non-cartesian CSes
take periodicity into account. The AABB is calculated in a way allowing to
show all coordinates used in data representation. E.g. if a segment
contains points with longitudes -170 and 170 then the AABB will be
[-180, 180], not [170, 190] nor [-190, -170] because then it wouldn't be
possible to show all of the longitudes used in data representation.

Implement drawing of non-cartesian Point, Segment and AABB taking
periodicity into account. In the last case draw prime and anti-meridians
as well as equator and poles. For consistency draw X and Y axes in
cartesian CS.